labs/fargate-dev-workshop/ecs_development_workshop/ecs_dev_network_stack.py
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved. SPDX-License-Identifier: MIT-0

from aws_cdk import (
    aws_ec2 as ec2,
    core
)

class EcsDevNetworkStack(core.Stack):

    def __init__(self, scope: core.Construct, id: str, **kwargs) -> None:
        super().__init__(scope, id, **kwargs)

        # The code that defines your stack goes here
        self.vpc = ec2.Vpc(
            self, "MyVpc",
            max_azs=3
        )
