labs/fargate-dev-workshop/requirements.txt
aws-cdk.aws-cloudformation
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-ecr
aws-cdk.aws-ecr-assets
aws-cdk.aws-s3
aws-cdk.aws-s3-assets
aws_cdk.aws_ecs_patterns
aws_cdk.aws_logs

aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_autoscaling
aws_cdk.aws_elasticloadbalancingv2_targets
aws_cdk.custom-resources

aws_cdk.aws_cloudwatch

aws-cdk.aws-sqs
aws-cdk.aws-ssm

aws-cdk.core
aws-cdk.cx-api
aws-cdk.region-info
  
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-lambda
aws-cdk.core